+++ ntex-net/src/polling/io.rs
@@ -30,12 +30,18 @@ struct HandleWrapper(WeakStreamCtl);
 impl Handle for HandleWrapper {
     fn query(&self, id: any::TypeId) -> Option<Box<dyn any::Any>> {
         if id == any::TypeId::of::<types::PeerAddr>() {
-            let addr = self.0.with(|io| io.peer_addr().ok());
+            let addr = self.0.with_socket(|io| io.peer_addr().ok());
             if let Some(addr) = addr.and_then(|addr| addr.as_socket()) {
                 return Some(Box::new(types::PeerAddr(addr)));
             }
         }
         None
+    }
+
+    fn write(&self, _: &IoContext) {
+        self.0.with(|item| {
+            item.write();
+        });
     }
 }